Refactor or relocation of various components
